[ue_wrap] an unresolved offset of -1 was written through, not refused
reflected_offset's accessors return -1 when the blueprint class has not loaded
or the field was renamed. Ten sites test for it; twenty-four did not, and the
worst is a write: WriteMainPlayerGrabbingPair stored two pointers at
base + (-1), one byte below the object, over its header. ReadMainPlayerGrabState
guarded holding_actor with a comment naming this exact hazard and read the four
fields beside it unguarded.

Fixed where the problem is, not per call site: puppet's ReadPtr/ReadAt/WriteAt
take a signed offset and refuse a negative one, which covers all sixteen uses in
that family, and the engine_mainplayer readers and writers now resolve the
offset into a named local and return early, the pattern the file already used
elsewhere. The size_t widening that turned -1 into a huge displacement is gone.

+++ src/votv-coop/src/ue_wrap/actors/puppet.cpp
@@ -83,7 +83,7 @@ bool ReadLookAtOnAnim(void* anim, FVector& out) {
     if (!IsKerfurAnimBP(anim)) return false;
     const int32_t off = ue_wrap::reflected_offset::AnimBP_kerfur_lookAt();
     if (off < 0) return false;
-    out = ReadAt<FVector>(anim, static_cast<size_t>(off));
+    out = ReadAt<FVector>(anim, off);
     return true;
 }
 void WriteLookAtOnAnim(void* anim, const FVector& target) {
@@ -91,8 +91,8 @@ void WriteLookAtOnAnim(void* anim, const FVector& target) {
     const int32_t lookOff   = ue_wrap::reflected_offset::AnimBP_kerfur_lookAt();
     const int32_t customOff = ue_wrap::reflected_offset::AnimBP_kerfur_customLookAt();
     if (lookOff < 0 || customOff < 0) return;
-    WriteAt<FVector>(anim, static_cast<size_t>(lookOff), target);
-    WriteAt<bool>(anim, static_cast<size_t>(customOff), true);
+    WriteAt<FVector>(anim, lookOff, target);
+    WriteAt<bool>(anim, customOff, true);
     static bool s_loggedDrive = false;
     if (!s_loggedDrive) { s_loggedDrive = true;
         UE_LOGI("puppet: kerfur head-look drive active -- wrote lookAt=(%.0f,%.0f,%.0f) + customLookAt=true (first; class-gate passed)",
